refactor(analysis): rewrite telemetry query pipeline with new record types - replace Map<byte[]> query with WeatherMapRecord-based repository port - refactor Kafka aggregation/deduplication processors and AggregationUtils - rename HASHTAG_DELIMITER → GEOHASH_SEPARATOR, update persistence adapters
refactor(history): restructure spatial query and telemetry listener
- relocate SpatialQueryKey under record package, refactor WeatherMapConverter
- update HistoricalTelemetryListener and add external communication observer IT

feat(ingestion): update telemetry controller and add dependencies

feat(common): add ServiceAddressRecord, utils, proto updates, sample resources

chore(infra): reorganize Jaeger config, update Consul/Vault/Kafka/deploy scripts

chore(tests): update JMeter stress tests and groovy scripts, add test resources

+++ history-service/src/main/java/me/neobliz1/ecomonitoring/platform/history/domain/model/dto/WeatherMapBucketCacheDto.java
@@ -6,6 +6,6 @@ public record WeatherMapBucketCacheDto(
         UUID id,
         long timestampBucket,
         int intervalMinutes,
-        int version
+        long version
 ) {
 }
